requirements.txt
django>=3.0.0
gunicorn
django-heroku
djangorestframework
django-filter